Implement new endpoint for filtering movies by filter ID
The update adds a new API endpoint in FilterController to filter movies according to a provided filter ID. The endpoint fetches the filter from the FilterService and applies it to the list of movies. Exception handlers are also added for cases when an invalid filter ID is provided or there are server issues.