Add backup_file.sql with database dump for application
This commit introduces the backup_file.sql in the project. This SQL file contains a dump of all the data and tables structure necessary for our application. It includes users, schema history, filters, criteria, and movies' data. This ensures that every instance of this app has an initial setup of database facilitating local and production deployments.